add games page
games page can be a modal, or a direct link to a game page.
Button to share the game with others
