+++ src/components/Modal/BoxScoreModal/index.js
@@ -1,1 +1,2 @@
 export * from './BoxScoreModal';
+export * from './BoxScoreContainer';
