+++ src/routes/Games/index.js
@@ -1,2 +1,3 @@
 export * from './EditGame';
 export * from './Game';
+export * from './BoxScoreData';
